Utilities

Utilities

JuAFEM.assemble!

JuAFEM.end_assemble

JuAFEM.function_divergence

JuAFEM.function_gradient

JuAFEM.function_symmetric_gradient

JuAFEM.function_value

JuAFEM.getcurrentface

JuAFEM.getdetJdV

JuAFEM.getdim

JuAFEM.getfacenumber

JuAFEM.getnbasefunctions

JuAFEM.getnquadpoints

JuAFEM.getorder

JuAFEM.getpoints

JuAFEM.getrefshape

JuAFEM.getweights

JuAFEM.reinit!

JuAFEM.shape_divergence

JuAFEM.shape_gradient

JuAFEM.shape_symmetric_gradient

JuAFEM.shape_value

JuAFEM.spatial_coordinate

JuAFEM.start_assemble

WriteVTK.vtk_grid

QuadratureRule

JuAFEM.getpoints — Function.

The points of the quadrature rule.

getpoints(qr::QuadratureRule)

Arguments:

qr: the quadrature rule

Example:

julia> qr = QuadratureRule{2, RefTetrahedron}(:legendre, 2);

julia> getpoints(qr)
3-element Array{Tensors.Tensor{1,2,Float64,2},1}:
 [0.166667, 0.166667]
 [0.166667, 0.666667]
 [0.666667, 0.166667]

source

JuAFEM.getweights — Function.

The weights of the quadrature rule.

getweights(qr::QuadratureRule) = qr.weights

Arguments:

qr: the quadrature rule

Example:

julia> qr = QuadratureRule{2, RefTetrahedron}(:legendre, 2);

julia> getweights(qr)
3-element Array{Float64,1}:
 0.166667
 0.166667
 0.166667

source

Interpolation

JuAFEM.getnbasefunctions — Function.

Returns the number of base functions for an Interpolation or Values object.

source

JuAFEM.getdim — Function.

Returns the dimension of an Interpolation

source

JuAFEM.getrefshape — Function.

Returns the reference shape of an Interpolation

source

JuAFEM.getorder — Function.

Returns the polynomial order of the Interpolation

source

CellValues

JuAFEM.reinit! — Function.

Updates a CellValues/FaceValues object for a cell or face.

reinit!{dim, T}(cv::CellValues{dim}, x::Vector{Vec{dim, T}})
reinit!{dim, T}(bv::FaceValues{dim}, x::Vector{Vec{dim, T}}, face::Int)

Arguments:

cv/bv: the CellValues/FaceValues object

x: a Vector of Vec, one for each nodal position in the element.

face: an integer to specify which face of the cell

Result

nothing

Details

source

JuAFEM.getnquadpoints — Function.

The number of quadrature points for the Values type.

getnquadpoints(fe_v::Values)

source

JuAFEM.getdetJdV — Function.

The product between the determinant of the Jacobian and the quadrature point weight for a given quadrature point: $\det(J(\mathbf{x})) w_q$

getdetJdV(fe_v::Values, quadrature_point::Int)

** Arguments:**

fe_v: the Values object

quadrature_point The quadrature point number

Results:

::Number

Details:

This value is typically used when one integrates a function on a finite element cell or face as

$\int\limits_\Omega f(\mathbf{x}) d \Omega \approx \sum\limits_{q = 1}^{n_q} f(\mathbf{x}_q) \det(J(\mathbf{x})) w_q$ $\int\limits_\Gamma f(\mathbf{x}) d \Gamma \approx \sum\limits_{q = 1}^{n_q} f(\mathbf{x}_q) \det(J(\mathbf{x})) w_q$

source

JuAFEM.shape_value — Function.

Computes the value of the shape function

shape_value(fe_v::Values, quadrature_point::Int, base_function::Int)

Gets the values of the shape function for a given quadrature point and base_func

source

JuAFEM.shape_gradient — Function.

Get the gradient of the shape functions for a given quadrature point and base function

source

JuAFEM.shape_symmetric_gradient — Function.

Get the symmetric gradient of the shape functions for a given quadrature point and base function

source

JuAFEM.shape_divergence — Function.

Get the divergence of the shape functions for a given quadrature point and base function

source

JuAFEM.function_value — Function.

Computes the value in a quadrature point for a scalar or vector valued function

function_value{dim, T}(fe_v::Values{dim}, q_point::Int, u::AbstractVector{T})
function_value{dim, T}(fe_v::Values{dim}, q_point::Int, u::AbstractVector{Vec{dim, T}})

Arguments:

fe_v: the Values object

q_point: the quadrature point number

u: the value of the function in the nodes

Results:

::Number: the value of a scalar valued function

::Vec{dim, T} the value of a vector valued function

Details:

The value of a scalar valued function is computed as $u(\mathbf{x}) = \sum\limits_{i = 1}^n N_i (\mathbf{x}) u_i$ where $u_i$ are the value of $u$ in the nodes. For a vector valued function the value is calculated as $\mathbf{u}(\mathbf{x}) = \sum\limits_{i = 1}^n N_i (\mathbf{x}) \mathbf{u}_i$ where $\mathbf{u}_i$ are the nodal values of $\mathbf{u}$.

source

JuAFEM.function_gradient — Function.

Computes the gradient in a quadrature point for a scalar or vector valued function

function_scalar_gradient{dim, T}(fe_v::Values{dim}, q_point::Int, u::AbstractVector{T})
function_vector_gradient{dim, T}(fe_v::Values{dim}, q_point::Int, u::AbstractVector{Vec{dim, T}})

Arguments:

fe_v: the Values object

q_point: the quadrature point number

u: the value of the function in the nodes

Results:

::Vec{dim, T}: the gradient of a scalar valued function

::Tensor{2, dim, T}: the gradient of a vector valued function

Details:

The gradient of a scalar function is computed as $\mathbf{\nabla} u(\mathbf{x}) = \sum\limits_{i = 1}^n \mathbf{\nabla} N_i (\mathbf{x}) u_i$ where $u_i$ are the nodal values of the function. For a vector valued function the gradient is computed as $\mathbf{\nabla} \mathbf{u}(\mathbf{x}) = \sum\limits_{i = 1}^n \mathbf{\nabla} N_i (\mathbf{x}) \otimes \mathbf{u}_i$ where $\mathbf{u}_i$ are the nodal values of $\mathbf{u}$.

source

JuAFEM.function_symmetric_gradient — Function.

Computes the symmetric gradient for a vector valued function in a quadrature point.

function_symmetric_gradient(fe_v::Values, q_point::Int, u::AbstractVector)

Arguments:

fe_v: the Values object

q_point: the quadrature point number

u: the value of the function in the nodes

Results:

::SymmetricTensor{2, dim, T}: the symmetric gradient

Details:

The symmetric gradient of a scalar function is computed as

$\left[ \mathbf{\nabla} \mathbf{u}(\mathbf{x_q}) \right]^\text{sym} = \sum\limits_{i = 1}^n \frac{1}{2} \left[ \mathbf{\nabla} N_i (\mathbf{x}_q) \otimes \mathbf{u}_i + \mathbf{u}_i \otimes \mathbf{\nabla} N_i (\mathbf{x}_q) \right]$

where $\mathbf{u}_i$ are the nodal values of the function.

source

JuAFEM.function_divergence — Function.

Computes the divergence in a quadrature point for a vector valued function.

function_divergence{dim, T}(fe_v::Values{dim}, q_point::Int, u::AbstractVector{Vec{dim, T}})

Arguments:

fe_v: the Values object

q_point: the quadrature point number

u: the value of the function in the nodes

Results:

::Number: the divergence of the function

Details:

The divergence of a vector valued functions in the quadrature point $\mathbf{x}_q)$ is computed as

$\mathbf{\nabla} \cdot \mathbf{u}(\mathbf{x_q}) = \sum\limits_{i = 1}^n \mathbf{\nabla} N_i (\mathbf{x_q}) \cdot \mathbf{u}_i$

where $\mathbf{u}_i$ are the nodal values of the function.

source

JuAFEM.spatial_coordinate — Function.

spatial_coordinate{dim, T}(fe_v::Values{dim}, q_point::Int, x::AbstractVector{Vec{dim, T}})

Computes the spatial coordinate in a quadrature point.

Arguments:

fe_v: the Values object

q_point: the quadrature point number

x: the nodal coordinates of the cell

Results:

::Vec{dim, T}: the spatial coordinate

Details:

The coordinate is computed, using the geometric interpolation, as $\mathbf{x} = \sum\limits_{i = 1}^n M_i (\mathbf{x}) \mathbf{\hat{x}}_i$

source

FaceValues

All of the methods for CellValues apply for FaceValues as well. In addition, there are some methods that are unique for FaecValues:

JuAFEM.getfacenumber — Function.

The face number for a cell, typically used to get the face number which is needed to reinit! a FaceValues object for face integration

getfacenumber(face_nodes, cell_nodes, ip::Interpolation)

** Arguments **

face_nodes: the node numbers of the nodes on the face of the cell

cell_nodes: the node numbers of the cell

ip: the Interpolation for the cell

** Results **

::Int: the corresponding face

source

JuAFEM.getcurrentface — Function.

The current active face of the FaceValues type.

getcurrentface(fv::FaceValues)

** Arguments **

fv: the FaceValues object

** Results **

::Int: the current active face (from last reinit!).

source

Assembling

JuAFEM.start_assemble — Function.

start_assemble([N=0]) -> Assembler

Call before starting an assembly.

Returns an Assembler type that is used to hold the intermediate data before an assembly is finished.

source

JuAFEM.assemble! — Function.

assemble!(a, Ke, edof)

Assembles the element matrix Ke into a.

source

assemble!(g, ge, edof)

Assembles the element residual ge into the global residual vector g.

source

JuAFEM.end_assemble — Function.

end_assemble(a::Assembler) -> K

Finalizes an assembly. Returns a sparse matrix with the assembled values.

source

VTK

WriteVTK.vtk_grid — Function.

vtk_grid(filename::AbstractString, grid::Grid)

Create a unstructured VTK grid from a Grid. Return a DatasetFile which data can be appended to, see vtk_point_data, vtk_cell_data.

source