FEValues

FEValues

CellValues

CellValues
reinit!
getnquadpoints
getdetJdV

shape_value
shape_gradient
shape_symmetric_gradient
shape_divergence

function_value
function_gradient
function_symmetric_gradient
function_divergence
spatial_coordinate

FaceValues

All of the methods for CellValues apply for FaceValues as well. In addition, there are some methods that are unique for FaecValues:

JuAFEM.FaceValues — Type.

FaceScalarValues([::Type{T}], quad_rule::QuadratureRule, func_interpol::Interpolation, [geom_interpol::Interpolation])
FaceVectorValues([::Type{T}], quad_rule::QuadratureRule, func_interpol::Interpolation, [geom_interpol::Interpolation])

A FaceValues object facilitates the process of evaluating values of shape functions, gradients of shape functions, values of nodal functions, gradients and divergences of nodal functions etc. on the faces of finite elements. There are two different types of FaceValues: FaceScalarValues and FaceVectorValues. As the names suggest, FaceScalarValues utilizes scalar shape functions and FaceVectorValues utilizes vectorial shape functions. For a scalar field, the FaceScalarValues type should be used. For vector field, both subtypes can be used.

Note

The quadrature rule for the face should be given with one dimension lower. I.e. for a 3D case, the quadrature rule should be in 2D.

Arguments:

T: an optional argument to determine the type the internal data is stored as.

quad_rule: an instance of a QuadratureRule

func_interpol: an instance of an Interpolation used to interpolate the approximated function

geom_interpol: an optional instance of an Interpolation which is used to interpolate the geometry

Common methods:

reinit!

getnquadpoints

getdetJdV

shape_value

shape_gradient

shape_symmetric_gradient

shape_divergence

function_value

function_gradient

function_symmetric_gradient

function_divergence

spatial_coordinate

source

JuAFEM.getcurrentface — Function.

getcurrentface(fv::FaceValues)

Return the current active face of the FaceValues object (from last reinit!).

source