Grid

JuAFEM.Node — Type

Node{dim, T}

A Node is a point in space.

Fields

x::Vec{dim,T}: stores the coordinates

JuAFEM.Cell — Type

Cell{dim,N,M} <: AbstractCell{dim,N,M}

A Cell is a sub-domain defined by a collection of Nodes as it's vertices. However, a cell is not defined by the nodes but rather by the node ids

Fields

nodes::Ntuple{N,Int}: N-tuple that stores the node ids

JuAFEM.Grid — Type

Grid{dim, C<:AbstractCell, T<:Real} <: AbstractGrid}

A Grid is a collection of Cells and Nodes which covers the computational domain, together with Sets of cells, nodes and faces.

Fields

cells::Vector{C}: stores all cells of the grid

nodes::Vector{Node{dim,T}}: stores the dim dimensional nodes of the grid

cellsets::Dict{String,Set{Int}}: maps a String key to a Set of cell ids

nodesets::Dict{String,Set{Int}}: maps a String key to a Set of node ids

facesets::Dict{String,Set{Tuple{Int,Int}}}: maps a String to a Set of Tuple{Int,Int} (global_cell_id, local_face_id)

boundary_matrix::SparseMatrixCSC{Bool,Int}: optional, only needed by onboundary to check if a cell is on the boundary, see, e.g. Helmholtz example

Utility Functions

JuAFEM.getcells — Function

getcells(grid::AbstractGrid) 
getcells(grid::AbstractGrid, v::Union{Int,Vector{Int}} 
getcells(grid::AbstractGrid, set::String)

Returns either all cells::Vector{C<:AbstractCell} of a grid or a subset based on an Int, Vector{Int} or String. Whereas the last option tries to call a cellset of the grid.

JuAFEM.getncells — Function

Returns and Int corresponding to how many cells are in the grid.

JuAFEM.getnodes — Function

getnodes(grid::AbstractGrid) 
getnodes(grid::AbstractGrid, v::Union{Int,Vector{Int}}
getnodes(grid::AbstractGrid, set::String)

Returns either all nodes::Vector{Node{dim,T}} of a grid or a subset based on an Int, Vector{Int} or String. The last option tries to call a nodeset of the grid.

JuAFEM.getnnodes — Function

returns an Int corresponding to how many nodes are in the grid

Missing docstring.

Missing docstring for nnodes_per_cell. Check Documenter's build log for details.

JuAFEM.getcellset — Function

Accesses the cellset which is mapped to the key set::String

JuAFEM.getnodeset — Function

Accesses the nodeset which is mapped to the key set::String

JuAFEM.getfaceset — Function

Accesses the faceset which is mapped to the key set::String

JuAFEM.compute_vertex_values — Function

function compute_vertex_values(grid::AbstractGrid, f::Function)
function compute_vertex_values(grid::AbstractGrid, v::Vector{Int}, f::Function)    
function compute_vertex_values(grid::AbstractGrid, set::String, f::Function)

Given a grid and some function f, compute_vertex_values computes all nodal values, i.e. values at the nodes, of the function f. The function implements two dispatches, where only a subset of the grid's node is used.

    compute_vertex_values(grid, x -> sin(x[1]) + cos([2]))
    compute_vertex_values(grid, [9, 6, 3], x -> sin(x[1]) + cos([2])) #compute function values at nodes with id 9,6,3
    compute_vertex_values(grid, "right", x -> sin(x[1]) + cos([2])) #compute function values at nodes belonging to nodeset right

JuAFEM.transform! — Function

transform!(grid::Abstractgrid, f::Function)

Transform all nodes of the grid based on some transformation function f.

JuAFEM.getcoordinates — Function

getcoordinates(grid::AbstractGrid, cell)

Return a vector with the coordinates of the vertices of cell number cell.

JuAFEM.getcoordinates! — Function

getcoordinates!(x::Vector{Vec{dim,T}}, grid::AbstractGrid, cell::Int)

Fills the vector x with the coordinates of a cell, defined by its cell id.

Grid Sets Utility

JuAFEM.addcellset! — Function

addcellset!(grid::AbstractGrid, name::String, cellid::Union{Set{Int}, Vector{Int}})
addcellset!(grid::AbstractGrid, name::String, f::function; all::Bool=true)

Adds a cellset::Dict{String,Set{Int}} to the grid with key name. Cellsets can be used to specify a boundary for Dirichlet, which is needed by the ConstraintHandler

addcellset!(grid, "left", Set((1,3))) #add cells with id 1 and 3 to cellset left
addcellset!(grid, "right", x -> norm(x[1]) < 2.0 ) #add cell to cellset right, if x[1] of each cell's node is smaller than 2.0

JuAFEM.addfaceset! — Function

addfaceset!(grid::AbstractGrid, name::String, faceid::Set{Tuple{Int,Int}})
addfaceset!(grid::AbstractGrid, name::String, f::Function; all::Bool=true)

Adds a faceset::Dict{String, Set{Tuple{Int,Int}} to the grid with key name. A faceset maps a String key to a Set of tuples corresponding to (global_cell_id, local_face_id). Facesets are used to initialize Dirichlet structs, that are needed to specify the boundary for the ConstraintHandler.

addfaceset!(gird, "right", Set(((2,2),(4,2))) #see grid manual example for reference
addfaceset!(grid, "clamped", x -> norm(x[1]) ≈ 0.0) #see incompressible elasticity example for reference

JuAFEM.addnodeset! — Function

addnodeset!(grid::AbstractGrid, name::String, nodeid::Union{Vector{Int},Set{Int}})
addnodeset!(grid::AbstractGrid, name::String, f::Function)

Adds a nodeset::Dict{String, Set{Int}} to the grid with key name. Has the same interface as addcellset. However, instead of mapping a cell id to the String key, a set of node ids is returned.