Main Types

JuAFEM.AbstractRefShape

JuAFEM.BoundaryValues

JuAFEM.CellValues

JuAFEM.Interpolation

JuAFEM.QuadratureRule

JuAFEM.AbstractRefShape — Type.

Represents a reference shape which quadrature rules and interpolations are defined on. Currently, the only concrete types that subtype this type are RefCube in 1,2 and 3 dimensions, and RefTetrahedron in 2 and 3 dimensions.

source

JuAFEM.QuadratureRule — Type.

A QuadratureRule is used to approximate an integral on a domain by a weighted sum of function values at specific points:

$\int\limits_\Omega f(\mathbf{x}) \text{d} \Omega \approx \sum\limits_{q = 1}^{n_q} f(\mathbf{x}_q) w_q$

The quadrature rule consists of $n_q$ points in space $\mathbf{x}_q$ with corresponding weights $w_q$.

There are different rules to determine the points and weights. In JuAFEM two different types are implemented: :legendre and :lobatto, where :lobatto is only supported for RefCube. If the quadrature rule type is left out, :legendre is used by default.

In JuAFEM, the QuadratureRule type is mostly used as one of the components to create a CellValues or BoundaryValues object.

Constructor:

QuadratureRule{dim, shape}([quad_rule_type::Symbol], order::Int)

Arguments:

dim: the space dimension of the reference shape

shape: an AbstractRefShape

quad_rule_type: :legendre or :lobatto, defaults to :legendre.

order: the order of the quadrature rule

Common methods:

getpoints : the points of the quadrature rule

getweights : the weights of the quadrature rule

Example:

julia> QuadratureRule{2, RefTetrahedron}(1)
JuAFEM.QuadratureRule{2,JuAFEM.RefTetrahedron,Float64}([0.5],Tensors.Tensor{1,2,Float64,2}[[0.333333,0.333333]])

julia> QuadratureRule{1, RefCube}(:lobatto, 2)
JuAFEM.QuadratureRule{1,JuAFEM.RefCube,Float64}([1.0,1.0],Tensors.Tensor{1,1,Float64,1}[[-1.0],[1.0]])

source

JuAFEM.Interpolation — Type.

An Interpolation is used to define shape functions to interpolate a function between nodes.

Constructor:

Interpolation{dim, reference_shape, order}()

Arguments:

dim: the dimension the interpolation lives in

shape: a reference shape, see AbstractRefShape

order: the highest order term in the polynomial

The following interpolations are implemented:

Lagrange{1, RefCube, 1}

Lagrange{1, RefCube, 2}

Lagrange{2, RefCube, 1}

Lagrange{2, RefCube, 2}

Lagrange{2, RefTetrahedron, 1}

Lagrange{2, RefTetrahedron, 2}

Lagrange{3, RefCube, 1}

Serendipity{2, RefCube, 2}

Lagrange{3, RefTetrahedron, 1}

Common methods:

getnbasefunctions

getdim

getrefshape

getorder

Example:

julia> ip = Lagrange{2, RefTetrahedron, 2}()
JuAFEM.Lagrange{2,JuAFEM.RefTetrahedron,2}()

julia> getnbasefunctions(ip)
6

source

JuAFEM.CellValues — Type.

A CellValues object facilitates the process of evaluating values of shape functions, gradients of shape functions, values of nodal functions, gradients and divergences of nodal functions etc. in the finite element cell. There are two different types of CellValues: CellScalarValues and CellVectorValues. As the names suggest, CellScalarValues utilizes scalar shape functions and CellVectorValues utilizes vectorial shape functions. For a scalar field, the CellScalarValues type should be used. For vector field, both subtypes can be used.

Constructors:

CellScalarValues([::Type{T}], quad_rule::QuadratureRule, func_interpol::Interpolation, [geom_interpol::Interpolation])
CellVectorValues([::Type{T}], quad_rule::QuadratureRule, func_interpol::Interpolation, [geom_interpol::Interpolation])

Arguments:

T: an optional argument to determine the type the internal data is stored as.

quad_rule: an instance of a QuadratureRule

func_interpol: an instance of an Interpolation used to interpolate the approximated function

geom_interpol: an optional instance of a Interpolation which is used to interpolate the geometry

Common methods:

reinit!

getnquadpoints

getquadrule

getfunctioninterpolation

getgeometryinterpolation

getdetJdV

shape_value

shape_gradient

shape_symmetric_gradient

shape_divergence

function_value

function_gradient

function_symmetric_gradient

function_divergence

spatial_coordinate

source

JuAFEM.BoundaryValues — Type.

A BoundaryValues object facilitates the process of evaluating values of shape functions, gradients of shape functions, values of nodal functions, gradients and divergences of nodal functions etc. on the finite element boundary. There are two different types of BoundaryValues: BoundaryScalarValues and BoundaryVectorValues. As the names suggest, BoundaryScalarValues utilizes scalar shape functions and BoundaryVectorValues utilizes vectorial shape functions. For a scalar field, the BoundaryScalarValues type should be used. For vector field, both subtypes can be used.

Constructors:

Note: The quadrature rule for the boundary should be given with one dimension lower. I.e. for a 3D case, the quadrature rule should be in 2D.

BoundaryScalarValues([::Type{T}], quad_rule::QuadratureRule, func_interpol::Interpolation, [geom_interpol::Interpolation])
BoundaryVectorValues([::Type{T}], quad_rule::QuadratureRule, func_interpol::Interpolation, [geom_interpol::Interpolation])

Arguments:

T: an optional argument to determine the type the internal data is stored as.

quad_rule: an instance of a QuadratureRule

func_interpol: an instance of an Interpolation used to interpolate the approximated function

geom_interpol: an optional instance of an Interpolation which is used to interpolate the geometry

Common methods:

reinit!

getboundarynumber

getnquadpoints

getquadrule

getfunctioninterpolation

getgeometryinterpolation

getdetJdV

shape_value

shape_gradient

shape_symmetric_gradient

shape_divergence

function_value

function_gradient

function_symmetric_gradient

function_divergence

spatial_coordinate

source