Utilities

Utilities

JuAFEM.assemble!

JuAFEM.end_assemble

JuAFEM.function_divergence

JuAFEM.function_gradient

JuAFEM.function_symmetric_gradient

JuAFEM.function_value

JuAFEM.getcurrentface

JuAFEM.getdetJdV

JuAFEM.getdim

JuAFEM.getnbasefunctions

JuAFEM.getnquadpoints

JuAFEM.getorder

JuAFEM.getpoints

JuAFEM.getrefshape

JuAFEM.getweights

JuAFEM.reinit!

JuAFEM.shape_divergence

JuAFEM.shape_gradient

JuAFEM.shape_symmetric_gradient

JuAFEM.shape_value

JuAFEM.spatial_coordinate

JuAFEM.start_assemble

WriteVTK.vtk_grid

QuadratureRule

JuAFEM.getpoints — Function.

getpoints(qr::QuadratureRule)

Return the points of the quadrature rule.

Examples

julia> qr = QuadratureRule{2, RefTetrahedron}(:legendre, 2);

julia> getpoints(qr)
3-element Array{Tensors.Tensor{1,2,Float64,2},1}:
 [0.166667, 0.166667]
 [0.166667, 0.666667]
 [0.666667, 0.166667]

source

JuAFEM.getweights — Function.

getweights(qr::QuadratureRule)

Return the weights of the quadrature rule.

Examples

julia> qr = QuadratureRule{2, RefTetrahedron}(:legendre, 2);

julia> getweights(qr)
3-element Array{Float64,1}:
 0.166667
 0.166667
 0.166667

source

Interpolation

JuAFEM.getnbasefunctions — Function.

Return the number of base functions for an Interpolation or Values object.

source

JuAFEM.getdim — Function.

Return the dimension of an Interpolation

source

JuAFEM.getrefshape — Function.

Return the reference shape of an Interpolation

source

JuAFEM.getorder — Function.

Return the polynomial order of the Interpolation

source

CellValues

JuAFEM.reinit! — Function.

reinit!(cv::CellValues, x::Vector)
reinit!(bv::FaceValues, x::Vector, face::Int)

Update the CellValues/FaceValues object for a cell or face with coordinates x. The derivatives of the shape functions, and the new integration weights are computed.

source

JuAFEM.getnquadpoints — Function.

getnquadpoints(fe_v::Values)

Return the number of quadrature points for the Values object.

source

JuAFEM.getdetJdV — Function.

getdetJdV(fe_v::Values, q_point::Int)

Return the product between the determinant of the Jacobian and the quadrature point weight for the given quadrature point: $\det(J(\mathbf{x})) w_q$

This value is typically used when one integrates a function on a finite element cell or face as

$\int\limits_\Omega f(\mathbf{x}) d \Omega \approx \sum\limits_{q = 1}^{n_q} f(\mathbf{x}_q) \det(J(\mathbf{x})) w_q$ $\int\limits_\Gamma f(\mathbf{x}) d \Gamma \approx \sum\limits_{q = 1}^{n_q} f(\mathbf{x}_q) \det(J(\mathbf{x})) w_q$

source

JuAFEM.shape_value — Function.

shape_value(fe_v::Values, q_point::Int, base_function::Int)

Return the value of shape function base_function evaluated in quadrature point q_point.

source

JuAFEM.shape_gradient — Function.

shape_gradient(fe_v::Values, q_point::Int, base_function::Int)

Return the gradient of shape function base_function evaluated in quadrature point q_point.

source

JuAFEM.shape_symmetric_gradient — Function.

shape_symmetric_gradient(fe_v::Values, q_point::Int, base_function::Int)

Return the symmetric gradient of shape function base_function evaluated in quadrature point q_point.

source

JuAFEM.shape_divergence — Function.

shape_divergence(fe_v::Values, q_point::Int, base_function::Int)

Return the divergence of shape function base_function evaluated in quadrature point q_point.

source

JuAFEM.function_value — Function.

function_value(fe_v::Values, q_point::Int, u::AbstractVector)

Compute the value of the function in a quadrature point. u is a vector with values for the degrees of freedom. For a scalar valued function, u contains scalars. For a vector valued function, u can be a vector of scalars (for use of VectorValues) or u can be a vector of Vecs (for use with ScalarValues).

The value of a scalar valued function is computed as $u(\mathbf{x}) = \sum\limits_{i = 1}^n N_i (\mathbf{x}) u_i$ where $u_i$ are the value of $u$ in the nodes. For a vector valued function the value is calculated as $\mathbf{u}(\mathbf{x}) = \sum\limits_{i = 1}^n N_i (\mathbf{x}) \mathbf{u}_i$ where $\mathbf{u}_i$ are the nodal values of $\mathbf{u}$.

source

JuAFEM.function_gradient — Function.

function_scalar_gradient(fe_v::Values{dim}, q_point::Int, u::AbstractVector)

Compute the gradient of the function in a quadrature point. u is a vector with values for the degrees of freedom. For a scalar valued function, u contains scalars. For a vector valued function, u can be a vector of scalars (for use of VectorValues) or u can be a vector of Vecs (for use with ScalarValues).

The gradient of a scalar function is computed as $\mathbf{\nabla} u(\mathbf{x}) = \sum\limits_{i = 1}^n \mathbf{\nabla} N_i (\mathbf{x}) u_i$ where $u_i$ are the nodal values of the function. For a vector valued function the gradient is computed as $\mathbf{\nabla} \mathbf{u}(\mathbf{x}) = \sum\limits_{i = 1}^n \mathbf{\nabla} N_i (\mathbf{x}) \otimes \mathbf{u}_i$ where $\mathbf{u}_i$ are the nodal values of $\mathbf{u}$.

source

JuAFEM.function_symmetric_gradient — Function.

function_symmetric_gradient(fe_v::Values, q_point::Int, u::AbstractVector)

Compute the symmetric gradient of the function, see function_gradient. Return a SymmetricTensor.

The symmetric gradient of a scalar function is computed as $\left[ \mathbf{\nabla} \mathbf{u}(\mathbf{x_q}) \right]^\text{sym} = \sum\limits_{i = 1}^n \frac{1}{2} \left[ \mathbf{\nabla} N_i (\mathbf{x}_q) \otimes \mathbf{u}_i + \mathbf{u}_i \otimes \mathbf{\nabla} N_i (\mathbf{x}_q) \right]$ where $\mathbf{u}_i$ are the nodal values of the function.

source

JuAFEM.function_divergence — Function.

function_divergence(fe_v::Values, q_point::Int, u::AbstractVector)

Compute the divergence of the vector valued function in a quadrature point.

The divergence of a vector valued functions in the quadrature point $\mathbf{x}_q)$ is computed as $\mathbf{\nabla} \cdot \mathbf{u}(\mathbf{x_q}) = \sum\limits_{i = 1}^n \mathbf{\nabla} N_i (\mathbf{x_q}) \cdot \mathbf{u}_i$ where $\mathbf{u}_i$ are the nodal values of the function.

source

JuAFEM.spatial_coordinate — Function.

spatial_coordinate(fe_v::Values{dim}, q_point::Int, x::AbstractVector)

Compute the spatial coordinate in a quadrature point. x contains the nodal coordinates of the cell.

The coordinate is computed, using the geometric interpolation, as $\mathbf{x} = \sum\limits_{i = 1}^n M_i (\mathbf{x}) \mathbf{\hat{x}}_i$

source

FaceValues

All of the methods for CellValues apply for FaceValues as well. In addition, there are some methods that are unique for FaecValues:

JuAFEM.getcurrentface — Function.

getcurrentface(fv::FaceValues)

Return the current active face of the FaceValues object (from last reinit!).

source

Assembling

JuAFEM.start_assemble — Function.

start_assemble([N=0]) -> Assembler

Call before starting an assembly.

Returns an Assembler type that is used to hold the intermediate data before an assembly is finished.

source

JuAFEM.assemble! — Function.

assemble!(a, Ke, edof)

Assembles the element matrix Ke into a.

source

assemble!(g, ge, edof)

Assembles the element residual ge into the global residual vector g.

source

JuAFEM.end_assemble — Function.

end_assemble(a::Assembler) -> K

Finalizes an assembly. Returns a sparse matrix with the assembled values.

source

VTK

WriteVTK.vtk_grid — Function.

vtk_grid(filename::AbstractString, grid::Grid)

Create a unstructured VTK grid from a Grid. Return a DatasetFile which data can be appended to, see vtk_point_data and vtk_cell_data.

source