Utilities

JuAFEM.function_divergence

JuAFEM.function_gradient

JuAFEM.function_symmetric_gradient

JuAFEM.function_value

JuAFEM.getVTKtype

JuAFEM.getboundarynumber

JuAFEM.getcurrentboundary

JuAFEM.getdetJdV

JuAFEM.getdim

JuAFEM.getfunctioninterpolation

JuAFEM.getgeometryinterpolation

JuAFEM.getnbasefunctions

JuAFEM.getnquadpoints

JuAFEM.getorder

JuAFEM.getpoints

JuAFEM.getquadrule

JuAFEM.getrefshape

JuAFEM.getweights

JuAFEM.reinit!

JuAFEM.shape_divergence

JuAFEM.shape_gradient

JuAFEM.shape_symmetric_gradient

JuAFEM.shape_value

JuAFEM.spatial_coordinate

WriteVTK.vtk_grid

QuadratureRule

JuAFEM.getpoints — Function.

The points of the quadrature rule.

getpoints(qr::QuadratureRule)

Arguments:

qr: the quadrature rule

Example:

julia> qr = QuadratureRule{2, RefTetrahedron}(:legendre, 2);

julia> getpoints(qr)
3-element Array{ContMechTensors.Tensor{1,2,Float64,2},1}:
 [0.166667,0.166667]
 [0.166667,0.666667]
 [0.666667,0.166667]

source

JuAFEM.getweights — Function.

The weights of the quadrature rule.

getweights(qr::QuadratureRule) = qr.weights

Arguments:

qr: the quadrature rule

Example:

julia> qr = QuadratureRule{2, RefTetrahedron}(:legendre, 2);

julia> getweights(qr)
3-element Array{Float64,1}:
 0.166667
 0.166667
 0.166667

source

Interpolation

JuAFEM.getnbasefunctions — Function.

Returns the number of base functions for an Interpolation or Values object.

source

JuAFEM.getdim — Function.

Returns the dimension of an Interpolation

source

JuAFEM.getrefshape — Function.

Returns the reference shape of an Interpolation

source

JuAFEM.getorder — Function.

Returns the polynomial order of the Interpolation

source

CellValues

JuAFEM.reinit! — Function.

Updates a CellValues/BoundaryValues object for a cell or boundary.

reinit!{dim, T}(cv::CellValues{dim}, x::Vector{Vec{dim, T}})
reinit!{dim, T}(bv::BoundaryValues{dim}, x::Vector{Vec{dim, T}}, boundary::Int)

Arguments:

cv/bv: the CellValues/BoundaryValues object

x: a Vector of Vec, one for each nodal position in the element.

boundary: an integer to specify which boundary of the cell

Result

nothing

Details

source

JuAFEM.getnquadpoints — Function.

The number of quadrature points for the Values type.

getnquadpoints(fe_v::Values)

source

JuAFEM.getquadrule — Function.

The quadrature rule for the Values type.

getquadrule(fe_v::Values)

** Arguments **

fe_v: the Values object

** Results **

::QuadratureRule: the quadrature rule.

source

JuAFEM.getfunctioninterpolation — Function.

The function interpolation for the Values type.

getfunctioninterpolation(fe_v::Values)

Arguments

fe_v: the Values object

Results

::Interpolation: the function interpolation

source

JuAFEM.getgeometryinterpolation — Function.

The interpolation used for the geometry for the Values type.

getgeometryinterpolation(fe_v::Values)

Arguments

fe_v: the Values object

Results

::Interpolation: the geometry interpolation

source

JuAFEM.getdetJdV — Function.

The product between the determinant of the Jacobian and the quadrature point weight for a given quadrature point: $\det(J(\mathbf{x})) w_q$

getdetJdV(fe_v::Values, quadrature_point::Int)

** Arguments:**

fe_v: the Values object

quadrature_point The quadrature point number

Results:

::Number

Details:

This value is typically used when one integrates a function on a finite element cell or boundary as

$\int\limits_\Omega f(\mathbf{x}) d \Omega \approx \sum\limits_{q = 1}^{n_q} f(\mathbf{x}_q) \det(J(\mathbf{x})) w_q$ $\int\limits_\Gamma f(\mathbf{x}) d \Gamma \approx \sum\limits_{q = 1}^{n_q} f(\mathbf{x}_q) \det(J(\mathbf{x})) w_q$

source

JuAFEM.shape_value — Function.

Computes the value of the shape function

shape_value(fe_v::Values, quadrature_point::Int, base_function::Int)

Gets the values of the shape function for a given quadrature point and base_func

source

JuAFEM.shape_gradient — Function.

Get the gradient of the shape functions for a given quadrature point and base function

source

JuAFEM.shape_symmetric_gradient — Function.

Get the symmetric gradient of the shape functions for a given quadrature point and base function

source

JuAFEM.shape_divergence — Function.

Get the divergence of the shape functions for a given quadrature point and base function

source

JuAFEM.function_value — Function.

Computes the value in a quadrature point for a scalar or vector valued function

function_value{dim, T}(fe_v::Values{dim}, q_point::Int, u::Vector{T})
function_value{dim, T}(fe_v::Values{dim}, q_point::Int, u::Vector{Vec{dim, T}})

Arguments:

fe_v: the Values object

q_point: the quadrature point number

u: the value of the function in the nodes

Results:

::Number: the value of a scalar valued function

::Vec{dim, T} the value of a vector valued function

Details:

The value of a scalar valued function is computed as $u(\mathbf{x}) = \sum\limits_{i = 1}^n N_i (\mathbf{x}) u_i$ where $u_i$ are the value of $u$ in the nodes. For a vector valued function the value is calculated as $\mathbf{u}(\mathbf{x}) = \sum\limits_{i = 1}^n N_i (\mathbf{x}) \mathbf{u}_i$ where $\mathbf{u}_i$ are the nodal values of $\mathbf{u}$.

source

JuAFEM.function_gradient — Function.

Computes the gradient in a quadrature point for a scalar or vector valued function

function_scalar_gradient{dim, T}(fe_v::Values{dim}, q_point::Int, u::Vector{T})
function_vector_gradient{dim, T}(fe_v::Values{dim}, q_point::Int, u::Vector{Vec{dim, T}})

Arguments:

fe_v: the Values object

q_point: the quadrature point number

u: the value of the function in the nodes

Results:

::Vec{dim, T}: the gradient of a scalar valued function

::Tensor{2, dim, T}: the gradient of a vector valued function

Details:

The gradient of a scalar function is computed as $\mathbf{\nabla} u(\mathbf{x}) = \sum\limits_{i = 1}^n \mathbf{\nabla} N_i (\mathbf{x}) u_i$ where $u_i$ are the nodal values of the function. For a vector valued function the gradient is computed as $\mathbf{\nabla} \mathbf{u}(\mathbf{x}) = \sum\limits_{i = 1}^n \mathbf{\nabla} N_i (\mathbf{x}) \otimes \mathbf{u}_i$ where $\mathbf{u}_i$ are the nodal values of $\mathbf{u}$.

source

JuAFEM.function_symmetric_gradient — Function.

Computes the symmetric gradient for a vector valued function in a quadrature point.

function_symmetric_gradient{dim, T}(fe_v::Values{dim}, q_point::Int, u::Vector{Vec{dim, T}})

Arguments:

fe_v: the Values object

q_point: the quadrature point number

u: the value of the function in the nodes

Results:

::SymmetricTensor{2, dim, T}: the symmetric gradient

Details:

The symmetric gradient of a scalar function is computed as

$\left[ \mathbf{\nabla} \mathbf{u}(\mathbf{x_q}) \right]^\text{sym} = \sum\limits_{i = 1}^n \frac{1}{2} \left[ \mathbf{\nabla} N_i (\mathbf{x}_q) \otimes \mathbf{u}_i + \mathbf{u}_i \otimes \mathbf{\nabla} N_i (\mathbf{x}_q) \right]$

where $\mathbf{u}_i$ are the nodal values of the function.

source

JuAFEM.function_divergence — Function.

Computes the divergence in a quadrature point for a vector valued function.

function_divergence{dim, T}(fe_v::Values{dim}, q_point::Int, u::Vector{Vec{dim, T}})

Arguments:

fe_v: the Values object

q_point: the quadrature point number

u: the value of the function in the nodes

Results:

::Number: the divergence of the function

Details:

The divergence of a vector valued functions in the quadrature point $\mathbf{x}_q)$ is computed as

$\mathbf{\nabla} \cdot \mathbf{u}(\mathbf{x_q}) = \sum\limits_{i = 1}^n \mathbf{\nabla} N_i (\mathbf{x_q}) \cdot \mathbf{u}_i$

where $\mathbf{u}_i$ are the nodal values of the function.

source

JuAFEM.spatial_coordinate — Function.

spatial_coordinate{dim, T}(fe_v::Values{dim}, q_point::Int, x::Vector{Vec{dim, T}})

Computes the spatial coordinate in a quadrature point.

Arguments:

fe_v: the Values object

q_point: the quadrature point number

x: the nodal coordinates of the cell

Results:

::Vec{dim, T}: the spatial coordinate

Details:

The coordinate is computed, using the geometric interpolation, as $\mathbf{x} = \sum\limits_{i = 1}^n M_i (\mathbf{x}) \mathbf{\hat{x}}_i$

source

BoundaryValues

All of the methods for CellValues apply for BoundaryValues as well. In addition, there are some methods that are unique for BoundaryValues:

JuAFEM.getboundarynumber — Function.

The boundary number for a cell, typically used to get the boundary number which is needed to reinit! a BoundaryValues object for boundary integration

getboundarynumber(boundary_nodes, cell_nodes, ip::Interpolation)

** Arguments **

boundary_nodes: the node numbers of the nodes on the boundary of the cell

cell_nodes: the node numbers of the cell

ip: the Interpolation for the cell

** Results **

::Int: the corresponding boundary

source

JuAFEM.getcurrentboundary — Function.

The current active boundary of the BoundaryValues type.

getcurrentboundary(bv::BoundaryValues)

** Arguments **

bv: the BoundaryValues object

** Results **

::Int: the current active boundary (from last reinit!).

source

VTK

WriteVTK.vtk_grid — Function.

Creates an unstructured VTK grid from the element topology and coordinates.

vtk_grid{dim,T}(filename::AbstractString, coords::Vector{Vec{dim,T}}, topology::Matrix{Int}, celltype::VTKCellTypes.VTKCellType)

Arguments

filename: name (or path) of the file when it is saved to disk, eg filename = "myfile", or filename = "/results/myfile" to store it in the folder results

coords: a vector of the node coordinates

topology: a matrix where each column contains the nodes which connects the element

celltype: the definition of the celltype in the grid, see https://github.com/jipolanco/WriteVTK.jl#defining-cells

Results:

::DatasetFile

Example:

julia> coords = [Vec{2}((0.0,0.0)), Vec{2}((1.0,0.0)), Vec{2}((1.5,1.5)), Vec{2}((0.0,1.0))]
4-element Array{ContMechTensors.Tensor{1,2,Float64,2},1}:
 [0.0,0.0]
 [1.0,0.0]
 [1.5,1.5]
 [0.0,1.0]

julia> topology = [1 2 4; 2 3 4]'
3×2 Array{Int64,2}:
 1  2
 2  3
 4  4

julia> celltype = VTKCellTypes.VTK_TRIANGLE;

julia> vtkobj = vtk_grid("example", coords, topology, celltype);

julia> vtk_save(vtkobj)
1-element Array{String,1}:
 "example.vtu"

Details

This is a thin wrapper around the function vtk_grid from the WriteVTK package.

For information how to add cell data and point data to the resulting VTK object as well as how to write it to a file see https://github.com/jipolanco/WriteVTK.jl#generating-an-unstructured-vtk-file

source

JuAFEM.getVTKtype — Function.

Returns the VTKCellType corresponding to the input Interpolation

getVTKtype(ip::Interpolation)

Arguments

ip: The interpolation

Results:

::VTKCellType: The cell type, see https://github.com/jipolanco/WriteVTK.jl#generating-an-unstructured-vtk-file

Example:

julia> ip = Lagrange{2, RefCube, 1}()
JuAFEM.Lagrange{2,JuAFEM.RefCube,1}()

julia> getVTKtype(ip)
WriteVTK.VTKCellTypes.VTKCellType("VTK_QUAD",0x09,4)

source