Interpolation

Interpolation

JuAFEM.Interpolation — Type.

Interpolation{dim, ref_shape, order}()

Return an Interpolation of given dimension dim, reference shape (see see AbstractRefShape) ref_shape and order order. order corresponds to the highest order term in the polynomial. The interpolation is used to define shape functions to interpolate a function between nodes.

The following interpolations are implemented:

Lagrange{1,RefCube,1}

Lagrange{1,RefCube,2}

Lagrange{2,RefCube,1}

Lagrange{2,RefCube,2}

Lagrange{2,RefTetrahedron,1}

Lagrange{2,RefTetrahedron,2}

Lagrange{3,RefCube,1}

Serendipity{2,RefCube,2}

Lagrange{3,RefTetrahedron,1}

Lagrange{3,RefTetrahedron,2}

Examples

julia> ip = Lagrange{2,RefTetrahedron,2}()
JuAFEM.Lagrange{2,JuAFEM.RefTetrahedron,2}()

julia> getnbasefunctions(ip)
6

source

JuAFEM.getnbasefunctions — Function.

Return the number of base functions for an Interpolation or Values object.

source

JuAFEM.getdim — Function.

Return the dimension of an Interpolation

source

JuAFEM.getrefshape — Function.

Return the reference shape of an Interpolation

source

JuAFEM.getorder — Function.

Return the polynomial order of the Interpolation

source